Commented out wasm deps for compatibility
I think this is a cargo issue where it tries resolving versions before
eliminating conditional compilation per target.

Example:
On the local repo we use a specific version of js_sys that conflicts
with this crate's version, even though we don't use wasm locally,
therefore the dependency should be dropped and not version checked.